# openapi_schema.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Cinema ERP - Firebase Simples",
                 "description":  "Sistema funcional com PostgreSQL e Firebase",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/v1/locations":  {
                                            "get":  {
                                                        "summary":  "Get Locations",
                                                        "description":  "Listar locaÃ§Ãµes",
                                                        "operationId":  "get_locations_api_v1_locations_get",
                                                        "parameters":  [
                                                                           {
                                                                               "name":  "skip",
                                                                               "in":  "query",
                                                                               "required":  false,
                                                                               "schema":  {
                                                                                              "type":  "integer",
                                                                                              "default":  0,
                                                                                              "title":  "Skip"
                                                                                          }
                                                                           },
                                                                           {
                                                                               "name":  "limit",
                                                                               "in":  "query",
                                                                               "required":  false,
                                                                               "schema":  {
                                                                                              "type":  "integer",
                                                                                              "default":  100,
                                                                                              "title":  "Limit"
                                                                                          }
                                                                           },
                                                                           {
                                                                               "name":  "status",
                                                                               "in":  "query",
                                                                               "required":  false,
                                                                               "schema":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Status"
                                                                                          }
                                                                           },
                                                                           {
                                                                               "name":  "city",
                                                                               "in":  "query",
                                                                               "required":  false,
                                                                               "schema":  {
                                                                                              "type":  "string",
                                                                                              "title":  "City"
                                                                                          }
                                                                           },
                                                                           {
                                                                               "name":  "include",
                                                                               "in":  "query",
                                                                               "required":  false,
                                                                               "schema":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Include"
                                                                                          }
                                                                           },
                                                                           {
                                                                               "name":  "page",
                                                                               "in":  "query",
                                                                               "required":  false,
                                                                               "schema":  {
                                                                                              "type":  "integer",
                                                                                              "default":  1,
                                                                                              "title":  "Page"
                                                                                          }
                                                                           },
                                                                           {
                                                                               "name":  "page_size",
                                                                               "in":  "query",
                                                                               "required":  false,
                                                                               "schema":  {
                                                                                              "type":  "integer",
                                                                                              "default":  12,
                                                                                              "title":  "Page Size"
                                                                                          }
                                                                           }
                                                                       ],
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Successful Response",
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {

                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  },
                                                                          "422":  {
                                                                                      "description":  "Validation Error",
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  }
                                                                      }
                                                    },
                                            "post":  {
                                                         "summary":  "Create Location",
                                                         "description":  "Criar nova locaÃ§Ã£o",
                                                         "operationId":  "create_location_api_v1_locations_post",
                                                         "requestBody":  {
                                                                             "required":  true,
                                                                             "content":  {
                                                                                             "application/json":  {
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "object",
                                                                                                                                     "additionalProperties":  true,
                                                                                                                                     "title":  "Location Data"
                                                                                                                                 }
                                                                                                                  }
                                                                                         }
                                                                         },
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "Successful Response",
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {

                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                           "422":  {
                                                                                       "description":  "Validation Error",
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   }
                                                                       }
                                                     }
                                        },
                  "/api/v1/locations/{location_id}":  {
                                                          "get":  {
                                                                      "summary":  "Get Location",
                                                                      "description":  "Buscar locaÃ§Ã£o por ID",
                                                                      "operationId":  "get_location_api_v1_locations__location_id__get",
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "location_id",
                                                                                             "in":  "path",
                                                                                             "required":  true,
                                                                                             "schema":  {
                                                                                                            "type":  "integer",
                                                                                                            "title":  "Location Id"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "Successful Response",
                                                                                                    "content":  {
                                                                                                                    "application/json":  {
                                                                                                                                             "schema":  {

                                                                                                                                                        }
                                                                                                                                         }
                                                                                                                }
                                                                                                },
                                                                                        "422":  {
                                                                                                    "description":  "Validation Error",
                                                                                                    "content":  {
                                                                                                                    "application/json":  {
                                                                                                                                             "schema":  {
                                                                                                                                                            "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                        }
                                                                                                                                         }
                                                                                                                }
                                                                                                }
                                                                                    }
                                                                  }
                                                      },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   }
                   }
}